refactor(init): split repo migration helpers Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
