+++ tests/cases/general/update-no-op/test.toml
@@ -1,4 +1,4 @@
 [expected]
 args = "update"
 exit = 0
-stdout = "flint: repo lint migration is up to date\n"
+stdout = "  normalized [tools] in <REPO>/mise.toml\n"
