+++ tests/cases/general/update-obsolete-key/test.toml
@@ -6,6 +6,7 @@ stdout = "  replaced \"github:mvdan/sh\" → \"shfmt\"\n"
 [expected.files]
 "mise.toml" = '''
 [tools]
+
 # Linters
 shellcheck = "v0.11.0"
 shfmt = "v3.13.1"
